requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-icons==0.2.5
django-crispy-forms==1.11.1
django-elasticsearch-dsl==7.1.4
django-haystack==3.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
Pillow==8.1.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-icons==0.2.5
django-crispy-forms==1.11.1
django-elasticsearch-dsl==7.1.4
django-haystack==3.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
lxml==4.6.3
Pillow==8.1.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-ajax-selects==2.1.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-icons==0.2.5
django-crispy-forms==1.11.1
django-elasticsearch-dsl==7.1.4
django-haystack==3.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
lxml==4.6.3
Pillow==8.1.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
